Move tutorial status constant into backend src
