+++ backend/src/shared/tutorialStatus.js
@@ -1,0 +1,7 @@
+const TUTORIAL_STATUS = Object.freeze({
+  DRAFT: 'draft',
+  PUBLISHED: 'published',
+  ARCHIVED: 'archived',
+});
+
+module.exports = { TUTORIAL_STATUS };
